+++ pkg/bdb/migrate/mig004/migrate.go
@@ -1,12 +1,6 @@
 package mig004
 
 import (
-	"bytes"
-	"context"
-
-	"github.com/aserto-dev/azm/model"
-	v3 "github.com/aserto-dev/azm/v3"
-	dsm3 "github.com/aserto-dev/go-directory/aserto/directory/model/v3"
 	"github.com/aserto-dev/go-edge-ds/pkg/bdb"
 	"github.com/aserto-dev/go-edge-ds/pkg/bdb/migrate/mig"
 	"github.com/rs/zerolog"
@@ -25,7 +19,7 @@ var fnMap = []func(*zerolog.Logger, *bolt.DB, *bolt.DB) error{
 	mig.CreateBucket(bdb.SystemPath),
 
 	mig.CreateBucket(bdb.ManifestPath),
-	migrateModel(),
+	mig.MigrateModel,
 
 	mig.CreateBucket(bdb.ObjectsPath),
 	mig.CreateBucket(bdb.RelationsObjPath),
@@ -41,46 +35,4 @@ func Migrate(log *zerolog.Logger, roDB, rwDB *bolt.DB) error {
 	}
 	log.Info().Str("version", Version).Msg("FinishedMigration")
 	return nil
-}
-
-func migrateModel() func(*zerolog.Logger, *bolt.DB, *bolt.DB) error {
-	return func(log *zerolog.Logger, roDB *bolt.DB, rwDB *bolt.DB) error {
-		// skip when roDB is nil.
-		if roDB == nil {
-			log.Debug().Msg("SKIP MigrateModel")
-			return nil
-		}
-
-		ctx := context.Background()
-		m, err := loadModel(ctx, roDB)
-		if err != nil {
-			return err
-		}
-
-		if err := rwDB.Update(func(tx *bolt.Tx) error {
-			_, err := bdb.SetAny[model.Model](ctx, tx, bdb.ManifestPath, bdb.ModelKey, m)
-			return err
-		}); err != nil {
-			return err
-		}
-
-		return nil
-	}
-}
-
-func loadModel(ctx context.Context, roDB *bolt.DB) (*model.Model, error) {
-	var m *model.Model
-	if err := roDB.View(func(rtx *bolt.Tx) error {
-		manifestBody, err := bdb.Get[dsm3.Body](ctx, rtx, bdb.ManifestPath, bdb.BodyKey)
-		if err != nil {
-			return err
-		}
-
-		m, err = v3.Load(bytes.NewReader(manifestBody.Data))
-		return err
-	}); err != nil {
-		return m, err
-	}
-
-	return m, nil
 }
