+++ pkg/bdb/migrate/migrate.go
@@ -12,6 +12,7 @@ import (
 	"github.com/aserto-dev/go-edge-ds/pkg/bdb/migrate/mig003"
 	"github.com/aserto-dev/go-edge-ds/pkg/bdb/migrate/mig004"
 	"github.com/aserto-dev/go-edge-ds/pkg/bdb/migrate/mig005"
+	"github.com/aserto-dev/go-edge-ds/pkg/bdb/migrate/mig006"
 	"github.com/aserto-dev/go-edge-ds/pkg/fs"
 
 	"github.com/Masterminds/semver"
@@ -29,6 +30,7 @@ var migMap = map[string]Migration{
 	mig003.Version: mig003.Migrate,
 	mig004.Version: mig004.Migrate,
 	mig005.Version: mig005.Migrate,
+	mig006.Version: mig006.Migrate,
 }
 
 var (
